TOML::Tiny: Provide no_string_guessing
Ideally, if we read a TOML file and write it back out again, we get a
semantically equivalent TOML file.  A necessary part of this is to
avoid the string-contents-based type-guessing.

Here, we provide a new `no_string_guessing` parameter which disables
this format-guessing.  Instead, it uses internal Perl flags to
distinguish integers from strings.

Right now this new feature is not likely to work for things that came
from TOML::Tiny::Parser without inflate_* parameters, since those are
all strings.  But it now allows Perl code to generate the TOML it
wants.