Writer: Fix param handling
The Writer doesn't have any params right now, but we're going to add
one.

 * Have to_toml capture its remaining args, the params, into a hash.
 * Have _to_toml expect params hashref as its 2nd argument
 * Pass through $params in two call sites

The existing code in what is now _to_toml was introduced in
5489d0cf0fa, when there was only to_toml.  It was attempting to cope
with the the fact that to_toml wanted to take params as an open-ended
argument list, but the recursive calls wanted to use a single hashref.
It also seems to have had an off-by-one error in indexing @_ - the
author was presumably confused by the renumbering of @_ due to the
shift just before.

Since 79a5b71e9bab we have to_toml (public) and _to_toml (internal),
so we don't need to test with ref any more.  But to_toml does need to
capture the parameters.

Add a function prototype to _to_toml to ensure that we didn't miss any
sites where the $param isn't passwed through, and to avoid any such
mistakes creeping in in the future.

No overall functional change.